John: Fix bugs in the computation of Nx2 Contingency Table baser and column offsets.

+++ met/lib/vx_contable/vx_nx2.cc
@@ -465,7 +465,7 @@ return ( x );
 
 double Nx2ContingencyTable::baser() const {
 
-   return(event_col_total()/n());
+   return((double) event_col_total()/n());
 }
